Decouple vehicle metadata from wasm bundle
